refactor(assertion): refactor
